feat: implementado forma de validacao se o usuario sera excluido ou anonimizado.

+++ epApi/src/main/java/br/fai/ep/api/service/impl/UsuarioServiceImpl.java
@@ -40,6 +40,12 @@ public class UsuarioServiceImpl implements BaseService {
 
     @Override
     public boolean delete(final long id) {
+        final Usuario user = (Usuario) dao.readById(id);
+        if (user.isAutor() || user.isParceiro() || user.isAdministrador()) {
+            user.setAnonimo(true);
+            return dao.update(user);
+        }
+
         return dao.delete(id);
     }
